+++ src/main/java/com/teamsupercat/roupangbackend/controller/ProductController.java
@@ -1,10 +1,7 @@
 package com.teamsupercat.roupangbackend.controller;
 
 import com.teamsupercat.roupangbackend.common.ResponseDto;
-import com.teamsupercat.roupangbackend.dto.CustomUserDetail.CustomUserDetail;
 import com.teamsupercat.roupangbackend.dto.option.request.OptionRegisterRequest;
-import com.teamsupercat.roupangbackend.dto.product.AllProductsResponse;
-import com.teamsupercat.roupangbackend.dto.product.ProductCreateRequest;
 import com.teamsupercat.roupangbackend.dto.product.ProductResponse;
 import com.teamsupercat.roupangbackend.service.OptionService;
 import com.teamsupercat.roupangbackend.service.ProductCategoryService;
@@ -15,10 +12,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
-import java.io.IOException;
 import java.text.ParseException;
 import java.util.List;
 import java.util.Map;
@@ -34,30 +29,6 @@ public class ProductController {
     private final OptionService optionService;
     private final ProductCategoryService productCategoryService;
 
-    @ApiOperation(value= "판매자 등록", notes = "판매자로 등록하기")
-    @PostMapping("/seller/signup")
-    public ResponseDto<Object> resisterAsSeller(@AuthenticationPrincipal CustomUserDetail userDetails){
-
-        Integer memberId = userDetails.getMemberIdx();
-
-        productService.saveAsSeller(memberId);
-
-        return new ResponseDto<>(true, "판매자로 등록되었습니다.", null);
-    }
-
-
-    @ApiOperation(value= "판매 물품 등록 ", notes = "판매할 물품 등록하기")
-    @PostMapping("/seller/products/register")
-    public ResponseDto<Object> createProduct(@RequestBody ProductCreateRequest productCreateRequest, @AuthenticationPrincipal CustomUserDetail userDetails) throws IOException, ParseException {
-
-        Integer memberId = userDetails.getMemberIdx();
-
-        productService.createProduct(productCreateRequest, memberId);
-
-        return new ResponseDto<>(true, "물품 등록에 성공하였습니다.",null);
-
-    }
-
     @ApiOperation(value= "판매 물품 상세 조회 ", notes = "물품 상세 조회")
     @GetMapping("/products/{product_id}")
     public ResponseDto<Object> getProductOne(@PathVariable("product_id") Integer productId) throws ParseException {
@@ -67,42 +38,6 @@ public class ProductController {
         return new ResponseDto<>(true, "물품 조회에 성공하였습니다.", productResponse);
 
     }
-
-
-    @ApiOperation(value= "판매자의 판매 물품 리스트 조회 ", notes = "판매자의 판매 물품 리스트 조회")
-    @GetMapping("/seller/products")
-    public ResponseDto<Object> getProductsList(@RequestParam(value = "order", required = false) String order, Pageable pageable, @AuthenticationPrincipal CustomUserDetail userDetails){
-
-        Integer memberId = userDetails.getMemberIdx();
-
-        List<AllProductsResponse> productsList = productService.getProductsList(order, pageable, memberId);
-
-        return new ResponseDto<>(true, "판매자의 판매 물품 내역 조회에 성공하였습니다.", productsList);
-    }
-
-    @ApiOperation(value= "판매자의 판매 물품 수정 ", notes = "판매자의 판매 물품 수정")
-    @PatchMapping("/seller/products/{product_id}")
-    public ResponseDto<Object> updateProduct(@PathVariable("product_id") Integer productId,
-                                             @RequestBody ProductCreateRequest productCreateRequest,
-                                             @AuthenticationPrincipal CustomUserDetail userDetails) throws ParseException {
-
-        Integer memberId = userDetails.getMemberIdx();
-        productService.updateProduct(productId, memberId, productCreateRequest);
-
-        return new ResponseDto<>(true, "판매자의 판매 물품이 정상적으로 업데이트되었습니다.", null);
-    }
-
-    @ApiOperation(value= "판매자의 판매 물품 삭제 ", notes = "판매자의 판매 물품 삭제")
-    @DeleteMapping("/seller/products/{product_id}")
-    public ResponseDto<Object> deleteProduct(@PathVariable("product_id") Integer productId,
-                                             @AuthenticationPrincipal CustomUserDetail userDetails) throws ParseException {
-
-        Integer memberId = userDetails.getMemberIdx();
-        productService.deleteProduct(productId, memberId);
-
-        return new ResponseDto<>(true, "판매자의 판매 물품이 정상적으로 삭제되었습니다.", null);
-    }
-
